Tweaked lifecycle for SQL engine.
Since the SQL engine runs on the driver, lifecycle is different from other plugins.

We will use the Spark Execution Context to execute the prepareRun and onRunFinish methods of the lifecycle for the SQL engine plugin.

This ensures we have access to the pipeline Runtime arguments within the SQL Engine.