CDAP-18076: Unset certificate path in CConfiguration when launching TaskWorkerTwillApplication
Why:
TaskWorkerService is currently used to run user code in isolation and
cdap-security is not exposed to it. So it has no access to certificate.
Unset the path in the cConf for TaskWorkerService, so it will generate
a certifacte itself if SSL is enabled. This is fine for now as currently
there is no code path verify the certificate (e.g. always trust servers)