CDAP-18010: Remove unused http REST endpoint TaskWorkerHttpHandlerInternal
Why:
This REST endpoint was introduced with the intention for artifact inspector
to upload artifact jar to run inspection in task worker.

Changing the implementation to upload jar to HDFS/GCS and send the location
to task worker service for it to download the jar, thus this REST endpoint
is obsolete and no longer needed.