CDAP-18451 setup guice bindings for spark on k8s
Added a step in the SparkRuntimeContextProvider to create and
instantiate a MasterEnvironment to use for service discovery and
other guice bindings.

To do this, added to the Spark pod template during submission in
order to mount the podinfo directory that the KubeMasterEnvironment
expects. Also added a way to specify the name of the master
environment SPI to use if it exists.