CDAP-18535 localize configs in k8s preparer using configmap
Added an option to tell the KubeTwillPreparer to use a configmap
to localize config files instead of having the FileLocalizer pull
them from app-fabric.

This lowers the load on app-fabric, and also removes the need
for a distributed filesystem to place those temporary files on.

Also enhanced the RemotePluginFinder to pull the artifact from
app-fabric in situations where the local filesystem is being
used for storing artifacts.