[CDAP-18627] Add put,get,getLatest,delete LevelDBTableCore APIs
Why:
Introduce put,get,delete APIs that operate on values at max version,
so caller doesn't need to specify a version when it doesn't care.
Introduce getLatest which returns the value with highest version.
This can be used when caller doesn't know the version number of
the latest value, but want to retrive the latest value.

Introducing these APIs allows simplifying the logic on the caller side.
For instance, a follow-up PR will make preview store to use these APIs.