CDAP-19133 fix spark launch namespace
Fix spark submission so that it sets the kubernetes namespace to
the correct value. To do this, added a way to get configuration
properties in the MasterEnvironment interface that gets the spark
submit configs.

In order to populate the configs, the namespace configs are now
stored in the system arguments for a program run. These are
populated at the start of the run and also in the TetheringAgent
when it forwards the program start message. The namespace
lookup was moved from the DistributedProgramRunner to the new
locations so that it is consistently available in the options
regardless of the program runner implementation or tethering,
o more easily access them from the SparkProgramRunner.

Also add a downward mount so that pods will contain their
namespace in a file instead of only looking in the CDAP conf for
the namespace that CDAP lives in. This fixes a bug where pods
would think their namespace is the cdap namespace instead of the
namespace they are running in.