CDAP-19153 strip artifact scheme and host for tethering
This fixes a bug in tethering scenarios when one CDAP instance
has one scheme for location URI (like hdfs://) and the peered
instance has a different scheme (like gs://). In these scenarios,
the LocationFactory cannot convert the URI into a Location
because of the scheme mismatch.

To fix this, the ArtifactCache service used in tethered runs will
strip the scheme and host from the location URI so that the
Location object can be created. This is a hack until the more
complete solution of removing Location from the ArtifactDetail
class can be implemented.