CDAP-19078 extend format interface for schema detection
Adding a method to ValidatingInputFormat so that format plugins
have a way to detect schema given some input data. This removes
the need for format plugins to interact with various filesystems,
allowing them to focus purely on how to read data instead of
also needing to know how to find which data to process.