CDAP-18966 remove job cleanup thread in preview
Modified the KubeTwillRunnerService so that it checks the configs
to see if it is being used for preview, which means it does not
need to monitor kubernetes jobs or clean up old jobs.

This reduces the load on kubernetes and also removes the possibility
of odd error messages in certain race conditions when threads from
multiple containers try to clean up the same job.