[CDAP-19376] Explicitly Check for Existing Service In KubeDiscoveryService
Trying to create service when all IP addresses are assigned results in
unhandled error. This is possible in RBAC enabled private network environments
with /27 address space because there we create exactly 31 services.
After initial startup, if any pod which performs the KubeDiscoveryService restarts, it will keep crashing on this issue.
The taskworker pods restart quite often and face this issue the most.

Testing Done -
1. Create a RBAC enabled private network with /27 address space.
2. After a few hours, observe that ALL taskworker pods are in INIT/terminate loop.
3. Update the image for all those pods with this fix.
4. Observe that taskworker pods now come up successfully. No pod restarts after couple of hours.
5. Deleted other pods such as runtime and appfabric pods. Observed the pods come up successfully.