CDAP-19465 Replacing kill() functionality by clean()
Adding a new method called clean() which can be used in RemoteExecutionTwillController complete().
This is a no-op method for the Tethering use case whereas for other it is same as kill().
This will ensure that a kill() command is not sent on Job complete for the Tethering use case
to resolve CDAP-19465.