CDAP-19468 configure Spark to return after submitting the job
Configure Spark to avoid waiting for the job to complete in
SparkSubmit and rely on our own pod watcher to determine if the
job is complete. Since we already are watching the pod, this
reduces the load on Kubernetes. It also bypasses a Spark bug
where it can wait forever even after the job is complete.