(CDAP-19827) Fix a race condition in AbstractProgramController
The AbstractProgramController.addListener() can leave behind an executor thread that is blocked forever if thread interruption was issued to the caller thread of the addListener method. This is because a SynchronousQueue was used to coordinate between the executor thread and the caller thread. If the caller thread is interrupted, it will unblock the take method, hence subsequence call to the put method from the executor thread will get blocked forever.