CDAP-20041 ensure extra jvm opts are properly set
make sure app.program.opts are always set in spark configs regardless
of if it's local mode or not. Also fix the modified entrypoint so that
these properties are set for the driver java process, since it is
getting ignored by spark submit, since we have hijacked the main
method.