CDAP-20445 reduce unactionable logging
Removed a warning log in MetadataConsumerSubscriberService when
reading a message that is not about field lineage, since it is
expected behavior and not a problem.

Reduced the frequency of the warning message about nosql
performance. When running on nosql, it would fill up the logs,
making it difficult to find more valuable information.