CDAP-20504 remove duplicate calls for secure store
Enhanced the SecretManager and SecureStore APIs to have
separate methods to get the actual secret content and the secret
metadata. This matches the REST API, which already has separate
endpoints for the data and metadata, and removes the need to make
two calls whenever just one piece of information is needed.

This also increases the usability of SecretManager implementations
that use a system that has different quotas for fetching metadata
and secret content, like the GCP SecretManager.