CDAP-20624 make entity id hash code deterministic
Make the EntityType hashCode() method deterministic by using the
enum ordinal instead of the enum hash code, which will differ
across jvms.

This was causing issues when using ApplicationId hashcode to
determine the program status topic to write to. With a deterministic
one, the same topic will always be chosen across JVMs.