CDAP-20649 fix dynamic plugin registration with client launch mode
Fixed a bug that caused the plugin archive to not be localized
when dynamically registering a plugin in client launch mode.
Instead of updating the plugin archive only in the
DistributedProgramRunner, refactored the logic so that the plugin
archive is updated right when the app spec is regenerated.
This removes the hidden tight coupling between the two classes,
and also fixes issues when the DistributedProgramRunner is not used,
such as in client launch mode.